requirements.txt
APScheduler==3.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
